Bugfix/fix wildcard queries (#216)
* Simplify ast, get rid of bindings, pass all variables explicitly

Having a binding was making queries more complex to develop and test.
Without them, all variables are passed explicitly. Moreover, :path
should have no special casing, it has to be the same exact operation
as all the others.

* Make it possible to run any-level wildcard queries.

Previously, the wildcard query in form of `a.b.*` with 
existing paths of `a.b.c` and `a.b.d` would return only
`a.b.c`, shadowing `a.b.d`. Moving add-date and traversals
to the new AST helped to mitigate this situation.

* Add tests for the wildcard queries

* Add integration test for the engine

* Remove a single array level from the structure

* Fix and improve tests

* Make sure that combined traversals work, split traversals for single and multi-series

Queries need to work for several cases:

1. When we combine series, sum
2. When we map series, absolute
3. When we reduce series, derivative
4. Combination of any

Since (1) is different from all the others, in the way it combines
series and has to do (apply mapv ...), we have to make a separate
function for it. FWIW, reducer and mapper _might_ be safely removed
from there.

To handle (2) and (3) in simpler way, `traverse-each!` was added,
that handles operations on singlular series.